updated avatar id, refid, fileid to long, added someother helpers as …

+++ Source/Podio .NET/Models/ByLine.cs
@@ -14,7 +14,7 @@ namespace PodioAPI.Models
         public string AvatarType { get; set; }
 
         [JsonProperty(PropertyName = "avatar_id")]
-        public int? AvatarId { get; set; }
+        public long? AvatarId { get; set; }
 
         [JsonProperty(PropertyName = "image")]
         public Image Image { get; set; }
@@ -26,6 +26,6 @@ namespace PodioAPI.Models
         public string Url { get; set; }
 
         [JsonProperty(PropertyName = "avatar")]
-        public int? Avatar { get; set; }
+        public long? Avatar { get; set; }
     }
 }
